Moved text selection utils to td_ui.
